+++ cmd/forklift/dev/env/cmd.go
@@ -11,37 +11,35 @@ var Cmd = &cli.Command{
 	Usage:   "Facilitates development and maintenance of a Forklift environment",
 	Flags: []cli.Flag{
 		&cli.StringSliceFlag{
-			Name:    "repo",
-			Aliases: []string{"r"},
-			Usage: "Replaces version-locked repos from the cache with the corresponding repos in " +
+			Name:    "plt",
+			Aliases: []string{"pallets"},
+			Usage: "Replaces version-locked pallets from the cache with the corresponding pallets in " +
 				"the specified directory paths",
 		},
 	},
 	Subcommands: []*cli.Command{
 		{
-			Name:     "cache-repo",
-			Aliases:  []string{"c-r", "cache-repositories"},
+			Name:     "cache-plt",
+			Aliases:  []string{"cache-pallets"},
 			Category: "Use the environment",
-			Usage:    "Updates the cache with the repositories available in the development environment",
-			Action:   cacheRepoAction,
+			Usage:    "Updates the cache with the pallets available in the development environment",
+			Action:   cachePalletAction,
 		},
 		{
 			Name:     "cache-img",
-			Aliases:  []string{"c-i", "cache-images"},
+			Aliases:  []string{"cache-images"},
 			Category: "Use the environment",
 			Usage:    "Pre-downloads the Docker container images required by the development environment",
 			Action:   cacheImgAction,
 		},
 		{
 			Name:     "check",
-			Aliases:  []string{"c"},
 			Category: "Use the environment",
 			Usage:    "Checks whether the development environment's resource constraints are satisfied",
 			Action:   checkAction,
 		},
 		{
 			Name:     "plan",
-			Aliases:  []string{"p"},
 			Category: "Use the environment",
 			Usage: "Determines the changes needed to update the Docker Swarm to match the deployments " +
 				"specified by the local environment",
@@ -49,7 +47,6 @@ var Cmd = &cli.Command{
 		},
 		{
 			Name:     "apply",
-			Aliases:  []string{"a"},
 			Category: "Use the environment",
 			Usage: "Updates the Docker Swarm to match the deployments specified by the " +
 				"development environment",
@@ -57,36 +54,35 @@ var Cmd = &cli.Command{
 		},
 		{
 			Name:     "show",
-			Aliases:  []string{"s"},
 			Category: "Query the environment",
 			Usage:    "Describes the development environment",
 			Action:   showAction,
 		},
 		{
-			Name:     "ls-repo",
-			Aliases:  []string{"ls-r", "list-repositories"},
+			Name:     "ls-plt",
+			Aliases:  []string{"list-pallets"},
 			Category: "Query the environment",
-			Usage:    "Lists repositories specified by the environment",
-			Action:   lsRepoAction,
+			Usage:    "Lists pallets specified by the environment",
+			Action:   lsPalletAction,
 		},
 		{
-			Name:      "show-repo",
-			Aliases:   []string{"s-r", "show-repository"},
+			Name:      "show-plt",
+			Aliases:   []string{"show-pallet"},
 			Category:  "Query the environment",
-			Usage:     "Describes a repository available in the development environment",
-			ArgsUsage: "repository_path",
-			Action:    showRepoAction,
+			Usage:     "Describes a pallet available in the development environment",
+			ArgsUsage: "pallet_path",
+			Action:    showPalletAction,
 		},
 		{
 			Name:     "ls-pkg",
-			Aliases:  []string{"ls-p", "list-packages"},
+			Aliases:  []string{"list-packages"},
 			Category: "Query the environment",
 			Usage:    "Lists packages available in the development environment",
 			Action:   lsPkgAction,
 		},
 		{
 			Name:      "show-pkg",
-			Aliases:   []string{"s-p", "show-package"},
+			Aliases:   []string{"show-package"},
 			Category:  "Query the environment",
 			Usage:     "Describes a package available in the development environment",
 			ArgsUsage: "package_path",
@@ -94,41 +90,41 @@ var Cmd = &cli.Command{
 		},
 		{
 			Name:    "ls-depl",
-			Aliases: []string{"ls-d", "list-deployments"},
+			Aliases: []string{"list-deployments"},
 			Usage:   "Lists package deployments specified by the development environment",
 			Action:  lsDeplAction,
 		},
 		{
 			Name:      "show-depl",
-			Aliases:   []string{"s-d", "show-deployment"},
+			Aliases:   []string{"show-deployment"},
 			Category:  "Query the environment",
 			Usage:     "Describes a package deployment specified by the development environment",
 			ArgsUsage: "package_path",
 			Action:    showDeplAction,
 		},
 		{
-			Name:      "add-repo",
-			Aliases:   []string{"add-r", "add-repositories"},
+			Name:      "add-plt",
+			Aliases:   []string{"add-pallets"},
 			Category:  "Query the environment",
-			Usage:     "Adds repositories to the environment, tracking specified versions or branches",
-			ArgsUsage: "[pallet_repository_path@version_query]...",
-			Action:    addRepoAction,
+			Usage:     "Adds pallets to the environment, tracking specified versions or branches",
+			ArgsUsage: "[pallet_path@version_query]...",
+			Action:    addPalletAction,
 		},
-		// TODO: add an upgrade-repo action?
+		// TODO: add an upgrade-pallet action?
 		// {
-		// 	Name:      "rm-repo",
-		// 	Aliases:   []string{"rm-r", "remove-repositories},
+		// 	Name:      "rm-plt",
+		// 	Aliases:   []string{"remove-pallets},
 		// 	Category:  "Modify the environment",
-		// 	Usage:     "Removes a repository from the environment",
-		// 	ArgsUsage: "repository_path",
+		// 	Usage:     "Removes a pallet from the environment",
+		// 	ArgsUsage: "pallet_path",
 		// 	Action: func(c *cli.Context) error {
-		// 		fmt.Println("removing repository", c.Args().First())
+		// 		fmt.Println("removing pallet", c.Args().First())
 		// 		return nil
 		// 	},
 		// },
 		// {
 		// 	Name:  "add-depl",
-		// 	Aliases:   []string{"add-d, "add-deployments"},
+		// 	Aliases:   []string{"add-deployments"},
 		// 	Category:  "Modify the environment",
 		// 	Usage: "Adds a package deployment to the environment",
 		// 	Action: func(c *cli.Context) error {
@@ -138,7 +134,7 @@ var Cmd = &cli.Command{
 		// },
 		// {
 		// 	Name:      "rm-depl",
-		// 	Aliases:   []string{"rm-d", "remove-deployments"},
+		// 	Aliases:   []string{"remove-deployments"},
 		// 	Category:  "Modify the environment",
 		// 	Usage:     "Removes a package deployment from the environment",
 		// 	ArgsUsage: "package_path",